+++ yivi_core/lib/src/models/schemaless/schemaless_events.g.dart
@@ -164,7 +164,6 @@ Credential _$CredentialFromJson(Map<String, dynamic> json) => Credential(
       : LogoImage.fromJson(json['image'] as Map<String, dynamic>),
   issuanceDate: (json['issuance_date'] as num?)?.toInt(),
   expiryDate: (json['expiry_date'] as num?)?.toInt(),
-  displayIsFallback: json['display_is_fallback'] as bool? ?? false,
 );
 
 Map<String, dynamic> _$CredentialToJson(Credential instance) =>
@@ -185,5 +184,4 @@ Map<String, dynamic> _$CredentialToJson(Credential instance) =>
       'revoked': instance.revoked,
       'revocation_supported': instance.revocationSupported,
       'issue_url': instance.issueUrl,
-      'display_is_fallback': instance.displayIsFallback,
     };